refactor: correct the imports for  ViewEventDataEntry and EditEventDataEntry
