+++ src/core_modules/capture-core/components/Pages/ViewEvent/index.js
@@ -1,5 +1,5 @@
 // @flow
-export { actionTypes as editEventDataEntryActionTypes } from './EventDetailsSection/EditEventDataEntry';
+export { actionTypes as editEventDataEntryActionTypes } from '../../WidgetEventEdit/EditEventDataEntry';
 export { assigneeSectionActionTypes } from '../ViewEvent/RightColumn/AssigneeSection';
 
 export { ViewEventPage } from './ViewEventPage.container';
