+++ src/core_modules/capture-core/components/WidgetEventEdit/EditEventDataEntry/editEventDataEntry.actions.js
@@ -1,7 +1,7 @@
 // @flow
 
-import { actionCreator } from '../../../../../actions/actions.utils';
-import { effectMethods } from '../../../../../trackerOffline';
+import { actionCreator } from '../../../actions/actions.utils';
+import { effectMethods } from '../../../trackerOffline';
 
 export const batchActionTypes = {
     START_SAVE_EDIT_EVENT_DATA_ENTRY_BATCH: 'StartSaveEditEventDataEntryBatchForViewSingleEvent',
